fix(history): Use category mapping until real cats
For now, use the Cardigann category mapping and the assigned custom categories. The only
thing we'd get from assigning custom categories in the definition is the ability to
query on the custom categories without having to lookup what custom category IDs
Jackett/Prowlarr assigned and the ability to query across indexers. The cost for that is
abusing the Cardigann framework and complicated and fragile per-indexer conditionals in
each definition. Given that looking up assigned custom category IDs need be done only
once per definition change and that there are many other reasons we may end up needing
to query each indexer individually, this isn't worth it.