fix(history): Derive category from other row data
I was misunderstanding how the `case:` block works in Cardigann. Now that I understand
that the switches are CSS selectors applied to the rows, it greatly simplifies how to
derive a category from the row itself.

Unfortunately, this still won't work because multiple "categories" can apply to each row
which means if the case switch order yield a different category than one in
`.Query.Categories`, that row will be filtered out even though one of the
`.Query.Categories` *is* a valid category for that row.